v0.7.2 (#36)
[Enhancement] Fixed issues in the code which caused warnings in MSVC and clang compilers.

[Enhancement] Fixed errors in `get_heuristics_list` where for certain heuristics mode in older cuDNN versions, the heuristics list might be incorrect.

[Bug fixes] Fixed several test cases failing on unsupported GPUs to exit gracefully.

[Samples] Added a sample to showcase fp8 convolution forward in Nvidia Hopper GPUs. The sample also showcases post convolution book-keeping operations such as scaling and absolute maximum reduction.

[Samples] Added a sample which converts fp16 tensor to fp8 and performs transpose and absolute maximum reduction.

[Samples] Added a sample to demonstrate Max pooling operation including tensor index dump, necessary to speed up the backward pass.

[Samples] Added a sample to showcase the backward pooling operation.

+++ include/cudnn_frontend_Engine.h
@@ -298,7 +298,7 @@ class EngineBuilder_v8 {
         }
 
 
-        for (uint64_t i = 0; i < m_engine.bKnobs.size(); i++) {
+        for (size_t i = 0; i < m_engine.bKnobs.size(); i++) {
             m_engine.bKnobs[i] = make_shared_backend_pointer(CUDNN_BACKEND_KNOB_INFO_DESCRIPTOR);
             if (m_engine.bKnobs[i]->is_good() == false) {
                 status = m_engine.bKnobs[i]->get_status();
